refactor(tui): complete DRY modular structure for diff/view TUIs
- Extract DiffDetailView and DiffApp to tui/diff/ module
- Extract ViewDetailView and ViewApp to tui/view/ module
- Convert diff_tui.py and view_tui.py to re-export shims
- All 1269 tests passing

TUI package structure now complete:
  tui/
  ├── base.py          # BaseBrowserApp, DetailView, COMMON_CSS
  ├── widgets.py       # SearchInput, HelpScreen, JumpDialog, SaveDialog
  ├── catalog/         # CatalogBrowserApp
  ├── diff/            # DiffApp, DiffDetailView
  └── view/            # ViewApp, ViewDetailView

Backwards compatibility maintained via shim files.

Authored by: Aaron Lippold<lippold@gmail.com>
